+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "sodoken"
-version = "0.0.901-alpha"
+version = "0.0.902-alpha"
 authors = ["Holochain Core Dev Team <devcore@holochain.org>"]
 edition = "2021"
 description = "libsodium wrapper providing tokio safe memory secure api access."
@@ -13,19 +13,11 @@ repository = "https://github.com/holochain/sodoken"
 resolver = "2"
 
 [dependencies]
-libc = "0.2.153"
-libsodium-sys = { version = "1.20.5", package = "libsodium-sys-stable" }
-
-# legacy deps
-num_cpus = { version = "1.16.0", optional = true }
-once_cell = { version = "1.19.0", optional = true }
-one_err = { version = "0.0.8", optional = true }
-parking_lot = { version = "0.12.1", optional = true }
-tokio = { version = "1.37.0", features = ["sync", "rt"], optional = true }
+libc = "0.2.169"
+libsodium-sys = { version = "1.22.2", package = "libsodium-sys-stable" }
 
 [dev-dependencies]
 criterion = "0.5.1"
-tokio = { version = "1.37.0", features = ["full"] }
 
 [package.metadata.docs.rs]
 all-features = true
@@ -33,20 +25,14 @@ all-features = true
 [features]
 default = []
 
-# legacy async api
-legacy = [ "dep:num_cpus", "dep:once_cell", "dep:one_err", "dep:parking_lot", "dep:tokio" ]
-
 [[bench]]
 name = "blake2b"
 harness = false
-required-features = [ "legacy" ]
 
 [[bench]]
 name = "random"
 harness = false
-required-features = [ "legacy" ]
 
 [[bench]]
 name = "sign"
 harness = false
-required-features = [ "legacy" ]
