--- src/legacy/hash.rs
@@ -1,4 +1,0 @@
-//! Modules related to hashing.
-
-pub mod argon2id;
-pub mod blake2b;
