--- src/legacy/safe.rs
@@ -1,2 +1,0 @@
-pub(crate) mod s3buf;
-pub(crate) mod sodium;
